Split row stream types
