+++ jetro-core/src/io/document_rows.rs
@@ -1,7 +1,8 @@
-use super::stream_exec::{CompiledRowStream, RowStreamRowResult};
+use super::stream_exec::CompiledRowStream;
 use super::stream_plan::{
     lower_root_rows_query, RowStreamDirection, RowStreamPlan, RowStreamSourceKind,
 };
+use super::stream_types::RowStreamRowResult;
 use crate::data::value::Val;
 use crate::{EvalError, Jetro, JetroEngine};
 use std::borrow::Cow;
